requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dateutils==0.6.11
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-WTF==0.14.3
fpdf==1.7.2
gunicorn==20.0.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.25
numpy==1.19.1
openpyxl==3.0.5
protobuf==3.17.2
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
simplejson==3.17.2
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
